Update clinic.cpp
Hey everyone, you guys did an excellent job on implementing every features that we said we would and especially the security feature is really a nice touch. This commit intends to edit some stuffs like using a uniform notation (camel notation) through out our entire code and Yonatan on the last commit you forgot to declare and define a function for storing stuff record and I did and I have also made some layout changes to and also changed the error messages when opening a file.